requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
amqp==5.1.1
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==3.2.14
django-bootstrap4==22.1
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-phonenumber-field==6.3.0
django-python-telegram-bot-adapter==0.1.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
dnspython==2.2.1
docutils==0.18.1
eventlet==0.33.1
frozenlist==1.3.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.4
iniconfig==1.1.1
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mattermostdriver==7.3.2
mmpy-bot==2.0.2
multidict==6.0.2
packaging==21.3
pep517==0.12.0
phonenumbers==8.12.50
pip-tools==6.6.2
pluggy==1.0.0
prompt-toolkit==3.0.29
py==1.11.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyTelegramBotAPI==4.6.0
pytest==7.1.2
python-dateutil==2.8.2
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
requests==2.28.0
rfc3986==1.5.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
telegram-django==1.1.2
tomli==2.0.1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
